Make seed type to be Int32 so that it is compatible with TPUs.
Otherwise, we get the following error:

```
 (OpKernel was found, but attributes didn't match) Requested Attributes: T=DT_INT32, Tseed=DT_INT64, dtype=DT_FLOAT
	.  Registered:  device='TPU'; Tseed in [DT_INT32]; T in [DT_INT32, DT_INT64]; dtype in [DT_FLOAT, DT_BFLOAT16]
```

+++ Sources/TensorFlow/Layers/Core.swift
@@ -214,8 +214,8 @@ public extension Dense {
         inputSize: Int,
         outputSize: Int,
         activation: @escaping Activation = identity,
-        seed: (Int64, Int64) = (Int64.random(in: Int64.min..<Int64.max),
-                                Int64.random(in: Int64.min..<Int64.max))
+        seed: (Int32, Int32) = (Int32.random(in: Int32.min..<Int32.max),
+                                Int32.random(in: Int32.min..<Int32.max))
     ) {
         self.init(weight: Tensor(glorotUniform: [inputSize, outputSize],
                                  seed: seed),
